+++ src/components/form-provenance/configured-form-provenance.jsx
@@ -3,7 +3,7 @@ import _throttle from 'lodash/throttle';
 
 import FormProvenance from './form-provenance.component';
 import validate from './form-provenance-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'provenance',
