+++ src/components/form-provenance/form-provenance-validations.js
@@ -1,8 +1,8 @@
 import _get from 'lodash/get';
 
-import { validateMinTwoChars } from '../../../validation/validation';
+import { validateMinTwoChars } from '../../validation/validation';
 
-import TranslationsService from '../../../services/translations';
+import TranslationsService from '../../services/translations';
 
 const validate = values => {
   const errors = {};
