+++ src/components/form-provenance/form-provenance.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import FormProvenance from './form-provenance.component';
-import frequency from '../../../../test/fixtures/frequency';
-import provenance from '../../../../test/fixtures/provenance';
+import frequency from '../../../test/fixtures/frequency';
+import provenance from '../../../test/fixtures/provenance';
 
 let defaultProps;
 let wrapper;
